+++ apps/browser/src/agent-jobs/AgentJobData.ts
@@ -1,4 +1,4 @@
-export type EvoAgentJobData = {
+export type AgentJobData = {
   threadId: string;
   shouldStop: boolean;
 };
